move image update in workloadmeta collector
