+++ pkg/workloadmeta/types.go
@@ -930,15 +930,6 @@ func (i *ContainerImageMetadata) Merge(e Entity) error {
 		return fmt.Errorf("cannot merge ContainerImageMetadata with different kind %T", e)
 	}
 
-	// SBOMs are generated only once. However, when they are generated it is possible that
-	// not every RepoDigest and RepoTags are attached to the image. In that case, the SBOM
-	// will also miss metadata and will not be re-generated when new metadata is detected.
-	// Because this metadata is essential for processing, it is important to inject new metadata
-	// to the existing SBOM. Generating a new SBOM can be a more robust solution but can also be
-	// costly.
-	// Moreover we can't update the SBOM in the collector because it's not thread safe.
-	i.SBOM = updateSBOMRepoMetadata(otherImage.SBOM, otherImage.RepoTags, otherImage.RepoDigests)
-
 	return merge(i, otherImage)
 }
